Make default TypeVisitor immutable
