+++ src/Psalm/Internal/TypeVisitor/ContainsClassLikeVisitor.php
@@ -5,16 +5,15 @@ namespace Psalm\Internal\TypeVisitor;
 use Psalm\Type\Atomic\TClassConstant;
 use Psalm\Type\Atomic\TLiteralClassString;
 use Psalm\Type\Atomic\TNamedObject;
-use Psalm\Type\ImmutableTypeVisitor;
-use Psalm\Type\TypeNode;
 use Psalm\Type\TypeVisitor;
+use Psalm\Type\TypeNode;
 
 use function strtolower;
 
 /**
  * @internal
  */
-class ContainsClassLikeVisitor extends ImmutableTypeVisitor
+class ContainsClassLikeVisitor extends TypeVisitor
 {
     /**
      * @var lowercase-string
@@ -43,21 +42,21 @@ class ContainsClassLikeVisitor extends ImmutableTypeVisitor
         if ($type instanceof TNamedObject) {
             if (strtolower($type->value) === $this->fq_classlike_name) {
                 $this->contains_classlike = true;
-                return TypeVisitor::STOP_TRAVERSAL;
+                return self::STOP_TRAVERSAL;
             }
         }
 
         if ($type instanceof TClassConstant) {
             if (strtolower($type->fq_classlike_name) === $this->fq_classlike_name) {
                 $this->contains_classlike = true;
-                return TypeVisitor::STOP_TRAVERSAL;
+                return self::STOP_TRAVERSAL;
             }
         }
 
         if ($type instanceof TLiteralClassString) {
             if (strtolower($type->value) === $this->fq_classlike_name) {
                 $this->contains_classlike = true;
-                return TypeVisitor::STOP_TRAVERSAL;
+                return self::STOP_TRAVERSAL;
             }
         }
