+++ src/Psalm/Internal/TypeVisitor/TypeChecker.php
@@ -25,10 +25,10 @@ use Psalm\Type\Atomic\TGenericObject;
 use Psalm\Type\Atomic\TNamedObject;
 use Psalm\Type\Atomic\TResource;
 use Psalm\Type\Atomic\TTemplateParam;
-use Psalm\Type\ImmutableTypeVisitor;
+use Psalm\Type\TypeVisitor;
 use Psalm\Type\MutableUnion;
 use Psalm\Type\TypeNode;
-use Psalm\Type\TypeVisitor;
+use Psalm\Type\MutableTypeVisitor;
 use Psalm\Type\Union;
 use ReflectionProperty;
 
@@ -42,7 +42,7 @@ use function strtolower;
 /**
  * @internal
  */
-class TypeChecker extends ImmutableTypeVisitor
+class TypeChecker extends TypeVisitor
 {
     /**
      * @var StatementsSource
@@ -118,7 +118,7 @@ class TypeChecker extends ImmutableTypeVisitor
         }
 
         if ($type->checked) {
-            return TypeVisitor::DONT_TRAVERSE_CHILDREN;
+            return self::DONT_TRAVERSE_CHILDREN;
         }
 
         if ($type instanceof TNamedObject) {
