+++ src/Psalm/Type/MutableUnion.php
@@ -498,7 +498,7 @@ final class MutableUnion implements TypeNode, Stringable
         return $union;
     }
 
-    public static function visitMutable(TypeVisitor $visitor, &$node, bool $cloned): bool
+    public static function visitMutable(MutableTypeVisitor $visitor, &$node, bool $cloned): bool
     {
         $result = true;
         $changed = false;
